Implementa Swagger y documentacion de ms-sucursales

+++ ms_sucursales/src/main/java/com/arriendo/ms_sucursales/repository/SucursalRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SucursalRepository extends JpaRepository<Sucursal, Integer> {
+public interface SucursalRepository extends JpaRepository<Sucursal, Long> {
 
     @Query(value = """
-            SELECT *
+            SELECT * 
             FROM sucursal
             WHERE operativa = true
             ORDER BY nombre ASC
